farmdurecore/requirements.txt
numpy==1.24.*
polars==0.18.*
pyyaml==6.*